# Optimize array allocation to boost the proxy activation performance